RJQ: replace old reader
