--- data_samples/conll03_new/conll03.conll
@@ -1,66 +1,0 @@
--DOCSTART- -X- -X- O
-
-The DT B-NP O
-European NNP I-NP B-ORG
-Commission NNP I-NP I-ORG
-said VBD B-VP O
-on IN B-PP O
-Thursday NNP B-NP O
-it PRP B-NP O
-disagreed VBD B-VP O
-with IN B-PP O
-German JJ B-NP B-MISC
-advice NN I-NP O
-to TO B-PP O
-consumers NNS B-NP O
-to TO B-VP O
-shun VB I-VP O
-British JJ B-NP B-MISC
-lamb NN I-NP O
-until IN B-SBAR O
-scientists NNS B-NP O
-determine VBP B-VP O
-whether IN B-SBAR O
-mad JJ B-NP O
-cow NN I-NP O
-disease NN I-NP O
-can MD B-VP O
-be VB I-VP O
-transmitted VBN I-VP O
-to TO B-PP O
-sheep NN B-NP O
-. . O O
-
-
--DOCSTART- -X- -X- O
-
-The DT B-NP O
-European NNP I-NP B-ORG
-Commission NNP I-NP I-ORG
-said VBD B-VP O
-on IN B-PP O
-Thursday NNP B-NP O
-it PRP B-NP O
-disagreed VBD B-VP O
-with IN B-PP O
-German JJ B-NP B-MISC
-advice NN I-NP O
-to TO B-PP O
-consumers NNS B-NP O
-to TO B-VP O
-shun VB I-VP O
-British JJ B-NP B-MISC
-lamb NN I-NP O
-until IN B-SBAR O
-scientists NNS B-NP O
-determine VBP B-VP O
-whether IN B-SBAR O
-mad JJ B-NP O
-cow NN I-NP O
-disease NN I-NP O
-can MD B-VP O
-be VB I-VP O
-transmitted VBN I-VP O
-to TO B-PP O
-sheep NN B-NP O
-. . O O
